# learning-platform/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AuthTokensResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.MeResponse:
    properties:
      avatarUrl:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
  dto.RefreshRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.RegisterRequest:
    properties:
      displayName:
        type: string
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - displayName
    - email
    - password
    type: object
  dto.RegisterResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.VerifyEmailRequest:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
  dto.VerifyResponse:
    properties:
      message:
        type: string
    type: object
  response.ErrorMessage:
    properties:
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorMessage'
      success:
        example: false
        type: boolean
    type: object
  response.SuccessWrapper:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for Learning Platform — users, auth, tasks, topics
  title: Learning Platform API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Returns access + refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  $ref: '#/definitions/dto.AuthTokensResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/me:
    get:
      description: Returns user info for current token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  $ref: '#/definitions/dto.MeResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Returns new access + refresh tokens
      parameters:
      - description: Refresh token payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  $ref: '#/definitions/dto.AuthTokensResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Refresh JWT tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user and sends email verification code
      parameters:
      - description: Register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  $ref: '#/definitions/dto.RegisterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /auth/verify:
    post:
      consumes:
      - application/json
      description: Verifies user email via code
      parameters:
      - description: Verify email payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  $ref: '#/definitions/dto.VerifyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Verify email
      tags:
      - auth
...
swagger: "2.0"